+++ lib/src/server/response/no_content_response.dart
@@ -1,8 +1,8 @@
 import 'package:json_api/document.dart';
-import 'package:json_api/src/server/response/response.dart';
+import 'package:json_api/src/server/response/json_api_response.dart';
 import 'package:json_api/src/server/server_document_factory.dart';
 
-class NoContentResponse extends Response {
+class NoContentResponse extends JsonApiResponse {
   const NoContentResponse() : super(204);
 
   @override