+++ lib/src/server/response/see_other_response.dart
@@ -1,9 +1,9 @@
 import 'package:json_api/document.dart';
-import 'package:json_api/src/server/response/response.dart';
+import 'package:json_api/src/server/response/json_api_response.dart';
 import 'package:json_api/src/server/server_document_factory.dart';
 import 'package:json_api/url_design.dart';
 
-class SeeOtherResponse extends Response {
+class SeeOtherResponse extends JsonApiResponse {
   final Resource resource;
 
   SeeOtherResponse(this.resource) : super(303);
@@ -12,8 +12,8 @@ class SeeOtherResponse extends Response {
   Document buildDocument(ServerDocumentFactory builder, Uri self) => null;
 
   @override
-  Map<String, String> getHeaders(UrlFactory route) => {
-        ...super.getHeaders(route),
-        'Location': route.resource(resource.type, resource.id).toString()
+  Map<String, String> getHeaders(UrlFactory urlFactory) => {
+        ...super.getHeaders(urlFactory),
+        'Location': urlFactory.resource(resource.type, resource.id).toString()
       };
 }